backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
browser-cookie3==0.19.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cssselect==1.2.0
curl-cffi==0.5.7
fastapi==0.103.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.3
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
g4f==0.0.2.8
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
Js2Py==0.74
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.3
multidict==6.0.4
newspaper3k==0.2.8
nltk==3.8.1
Pillow==10.0.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
pyjsparser==2.7.1
python-dateutil==2.8.2
PyYAML==6.0.1
quickjs==1.19.2
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tinysegmenter==0.3
tldextract==3.5.0
tqdm==4.66.1
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.4
uvicorn==0.23.2
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.2
